dom/tests/mochitest/chrome/test_bug830396.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=830396
-->
<window title="Mozilla Bug 830396"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <title>Test for Bug 830396</title>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
<script class="testbody" type="application/javascript">
<![CDATA[

function runTests()
{
  var doc = frames[0].document;
  var cs = doc.defaultView.getComputedStyle(doc.body);
  var nsIDOMCSSValueList = Components.interfaces.nsIDOMCSSValueList;
  ok(cs.getPropertyCSSValue("cursor") instanceof nsIDOMCSSValueList,
     "CSSValueList should be a nsIDOMCSSValueList");
  SimpleTest.finish();
}
]]>
</script>
<body onload="runTests();"  xmlns="http://www.w3.org/1999/xhtml">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=830396">Mozilla Bug 830396</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
</pre>
<iframe type="content"></iframe>
</body>


</window>
